Pulled some of the create connection inputs out to their own file
