+++ libs/designer/src/lib/ui/panel/index.tsx
@@ -1,2 +1,2 @@
-export * from './connectionsPanel/createConnection/universalConnectionParameter';
+export * from './connectionsPanel/createConnection/formInputs/universalConnectionParameter';
 export * from './connectionsPanel/createConnection/connectionParameterRow';
